tests/requirements-test.txt
PyYAML>=5.1.1
pytest>=4.6.3
jsonschema>=3.0.1
rfc3987>=1.3.8
webcolors>=1.9.1
strict-rfc3339>=0.7
idna>=2.8
jsonpointer>=2.0
